requirements.txt
streamlit>=1.32
google-genai>=0.6.0
google-auth>=2.20.0
